feat(refactor): Move common exceptions out of the discover db
Move IssueSearchQueryException to common exceptions module

Remove future annotation

Remove future annotation
