+++ src/sentry/search/events/datasets/discover.py
@@ -95,21 +95,11 @@ from sentry.search.events.fields import (
 )
 from sentry.search.events.filter import to_list
 from sentry.search.events.types import SelectType, WhereType
+from sentry.search.exceptions import InvalidIssueSearchQuery
 from sentry.search.utils import DEVICE_CLASS
 from sentry.snuba.dataset import Dataset
 from sentry.snuba.referrer import Referrer
 from sentry.utils.numbers import format_grouped_length
-
-
-class InvalidIssueSearchQuery(InvalidSearchQuery):
-    """Raised when an issue filter references non-existent issue IDs."""
-
-    def __init__(self, invalid_ids: list[str]):
-        self.invalid_ids = invalid_ids
-        super().__init__(f"Issue IDs do not exist: {invalid_ids}")
-
-    def __str__(self) -> str:
-        return f"Issue IDs do not exist: {self.invalid_ids}"
 
 
 class DiscoverDatasetConfig(DatasetConfig):
